Embeding, vector database ve semantic search/04-LLM/requirements.txt
# Kararlı ve uyumlu olduğu bilinen kütüphane sürümleri
chromadb==0.4.24
openai==1.12.0
pandas==2.1.4
python-dotenv==1.0.0
pydantic<2
numpy==1.26.2